remove http_host from configuration files
The API and configuration files changed and do not include "http_host"
anymore. Therefore it does not need to be set for the test application
and the crm will deduce the value from its url, because the url schema
now is "https://project.crm.infopark.net" or
"https://project.cms.infopark.net" respectively.

Until recently, only a placeholder for the actual project was part of
the configuration urls. This "admin" part is now not longer needed as
the urls contain the correct project name.