add dynamic page title
Each page has a title that is displayed in the head section of the DOM.
If a page did not define a title itself, then its homepage title or the
title of the default homepage is used. The homepage title is also
appended to all subpage titles. The NullHomepage represents a special
case where a static title is set.

A title cell renders the section in the main layout. This cell will be
extended in the future to include other seo relevant information like
keywords or the description.

Currently, we are using the system title attribute of the CMS. In the
future this should become a local attribute "title".

fixes #75